+++ libraries/base/changelog.md
@@ -11,6 +11,8 @@
   * Introduce `GHC.ExecutablePath.executablePath`, which is more robust than
     `getExecutablePath` in cases when the executable has been deleted.
 
+  * Add `Data.ByteArray` module, providing a boxed `ByteArray#`.
+
 ## 4.16.0.0 *TBA*
 
   * Make it possible to promote `Natural`s and remove the separate `Nat` kind.